Added the feature to export model-image intesection areas as binary files

+++ src/CAPMath.h
@@ -33,7 +33,7 @@ public:
 	}
 
 	Real Length() const {
-		return sqrtf(x * x + y * y + z * z);
+		return sqrt(x * x + y * y + z * z);
 	}
 
 	void Normalise() {
